added category service, TODO propagation of creation of products and categories to the other service

+++ productmanagement/src/main/java/com/hka/verteiltesysteme/controller/ProductController.java
@@ -18,7 +18,7 @@ public class ProductController {
     @PostMapping("/product")
     public ResponseEntity<String> CreateProduct(@RequestBody UpsertProductDto createProductDto) {
         try {
-            Category category = categoryRepo.getById((int) createProductDto.categoryId());
+            Category category = categoryRepo.getById(createProductDto.categoryId());
             var product = Product.builder().name(createProductDto.name())
                     .category(category)
                     .price(createProductDto.price())
